added checks to the boot192.sh file to avoid adding stuff twice

+++ boot192.sh
@@ -41,21 +41,27 @@ apt-get update
 #### installing Ruby 1.9.2
 
 # install Ruby 1.9.2 from source
-cd /tmp
-wget ftp://ftp.ruby-lang.org//pub/ruby/1.9/ruby-1.9.2-p0.tar.gz
-tar -xvzf ruby-1.9.2-p0.tar.gz
-cd ruby-1.9.2-p0/
-./configure --prefix=/usr/local/ruby
-make && sudo make install
-
-# add ruby to the PATH
-sed -e '/^PATH/s/"$/:\/usr\/local\/ruby\/bin"/g' -i /etc/environment
-#source /etc/environment
+if [ ! -e "/usr/local/ruby" && ! command -v ruby ]
+  cd /tmp
+  wget ftp://ftp.ruby-lang.org//pub/ruby/1.9/ruby-1.9.2-p0.tar.gz
+  tar -xvzf ruby-1.9.2-p0.tar.gz
+  cd ruby-1.9.2-p0/
+  ./configure --prefix=/usr/local/ruby
+  make && sudo make install
+  # add ruby to the PATH
+  sed -e '/^PATH/s/"$/:\/usr\/local\/ruby\/bin"/g' -i /etc/environment
+  #source /etc/environment
+fi
 
 # set symbolic links
-rm /usr/local/bin/ruby
+if [ -e "/usr/local/bin/ruby" ]
+  rm /usr/local/bin/ruby
+fi
 ln -s /usr/local/ruby/bin/ruby /usr/local/bin/ruby
-rm /usr/bin/gem
+
+if [ -e "/usr/bin/gem" ]
+  rm /usr/bin/gem
+fi
 ln -s /usr/local/ruby/bin/gem /usr/bin/gem
 
 # install bundler
@@ -71,6 +77,9 @@ gem install ohai --no-rdoc --no-ri
 gem install chef --no-rdoc --no-ri
   
 # clone unepwcmc chef repo
+if [ -d /tmp/cookbooks ]
+  rm -rf /tmp/cookbooks
+fi
 cd /tmp
 git clone http://github.com/unepwcmc/cookbooks.git
 cd /tmp/cookbooks
